+++ src/optimizers/optimizers.cu
@@ -1,15 +1,17 @@
 #include "optimizers.h"
 
 #include "tensors/tensor_operators.h"
-#include "functional/functional.h"
 
 namespace marian {
+
 void Sgd::updateImpl(Tensor params, Tensor grads) {
   using namespace functional;
   Element(_1 -= (multiplyFactor_ * eta_) * _2, params, grads);
 
   //cudaStreamSynchronize(0);
 }
+
+// Aagrad
 
 void Adagrad::updateImpl(Tensor params, Tensor grads) {
   if(!alloc_)
@@ -34,11 +36,93 @@ void Adagrad::updateImpl(Tensor params, Tensor grads) {
   //cudaStreamSynchronize(0);
 }
 
+void Adagrad::load(const std::string& name,
+                   std::vector<Ptr<OptimizerBase>> opts,
+                   std::vector<Ptr<Backend>> backends) {
+  if(!boost::filesystem::exists(name))
+    return;
+
+  LOG(info, "Loading Adagrad parameters from {}", name);
+
+  std::vector<float> vGt;
+  size_t totalSize = 0;
+
+  auto numpy = cnpy::npz_load(name);
+  for(auto it : numpy) {
+    auto name = it.first;
+    cnpy::NpyArray& np = it.second;
+
+    // get the size of gt_
+    totalSize = np.shape[1];
+
+    // extract data into vectors
+    if(name == "adagrad_gt") {
+      vGt.resize(totalSize);
+      std::copy((float*)np.data, (float*)np.data + totalSize, vGt.begin());
+    }
+  }
+
+  if(vGt.empty()) {
+    LOG(warn, "[warn] Adagrad parameters not found in .npz file");
+    return;
+  }
+
+  // get the size of params which should go
+  size_t shardSize = ceil(totalSize / (float)backends.size());
+
+  size_t id = 0;
+  for(auto optBase : opts) {
+    auto opt = std::dynamic_pointer_cast<Adagrad>(optBase);
+
+    int size = std::min(shardSize, totalSize);
+    totalSize -= size;
+
+    if(!opt->alloc_)
+      opt->alloc_ = New<TensorAllocator>(backends[id]);
+
+    if(!opt->gt_) {
+      opt->alloc_->reserveExact(sizeof(float) * size);
+      opt->alloc_->allocate(opt->gt_, {1, size});
+    }
+
+    size_t shift = id * shardSize;
+    std::vector<float> tmp(vGt.begin() + shift, vGt.begin() + shift + size);
+    opt->gt_->set(tmp);
+
+    id++;
+  }
+}
+
+void Adagrad::save(const std::string& name,
+                   std::vector<Ptr<OptimizerBase>> opts,
+                   size_t totalSize) {
+  LOG(info, "Saving Adagrad parameters to {}", name);
+
+  std::vector<float> vGt;
+
+  for(auto optBase : opts) {
+    auto opt = std::dynamic_pointer_cast<Adagrad>(optBase);
+    std::vector<float> tmp;
+    opt->gt_->get(tmp);
+    vGt.insert(vGt.end(), tmp.begin(), tmp.end());
+  }
+
+  unsigned* shape = new unsigned[2];
+  shape[0] = 1;
+  shape[1] = vGt.size();
+
+  cnpy::npz_save(name, "adagrad_gt", vGt.data(), shape, 2, "w");
+
+  delete[] shape;
+}
+
 void Adagrad::resetStats() {
   if(gt_)
     gt_->set(0);
   //cudaStreamSynchronize(0);
 }
+
+// Adam
 
 void Adam::updateImpl(Tensor params, Tensor grads) {
   if(!alloc_)
@@ -70,6 +154,100 @@ void Adam::updateImpl(Tensor params, Tensor grads) {
           vt_);
 
   //cudaStreamSynchronize(0);
+}
+
+void Adam::load(const std::string& name,
+                std::vector<Ptr<OptimizerBase>> opts,
+                std::vector<Ptr<Backend>> backends) {
+  if(!boost::filesystem::exists(name))
+    return;
+
+  LOG(info, "Loading Adam parameters from {}", name);
+
+  std::vector<float> vMt;
+  std::vector<float> vVt;
+  size_t totalSize = 0;
+
+  auto numpy = cnpy::npz_load(name);
+  for(auto it : numpy) {
+    auto name = it.first;
+    cnpy::NpyArray& np = it.second;
+
+    // get the size of mt_ and vt_, they are the same
+    totalSize = np.shape[1];
+
+    // extract data into vectors
+    if(name == "adam_mt") {
+      vMt.resize(totalSize);
+      std::copy((float*)np.data, (float*)np.data + totalSize, vMt.begin());
+    }
+    if(name == "adam_vt") {
+      vVt.resize(totalSize);
+      std::copy((float*)np.data, (float*)np.data + totalSize, vVt.begin());
+    }
+  }
+
+  if(vMt.empty() || vVt.empty()) {
+    LOG(warn, "[warn] Adam parameters not found in .npz file");
+    return;
+  }
+
+  // get the size of params which should go
+  size_t shardSize = ceil(totalSize / (float)backends.size());
+
+  size_t id = 0;
+  for(auto optBase : opts) {
+    auto opt = std::dynamic_pointer_cast<Adam>(optBase);
+
+    int size = std::min(shardSize, totalSize);
+    totalSize -= size;
+
+    if(!opt->alloc_)
+      opt->alloc_ = New<TensorAllocator>(backends[id]);
+
+    if(!opt->mt_ || !opt->vt_) {
+      opt->alloc_->reserveExact(2 * sizeof(float) * size);
+      opt->alloc_->allocate(opt->mt_, {1, size});
+      opt->alloc_->allocate(opt->vt_, {1, size});
+    }
+
+    size_t shift = id * shardSize;
+    std::vector<float> tmpMt(vMt.begin() + shift, vMt.begin() + shift + size);
+    opt->mt_->set(tmpMt);
+    std::vector<float> tmpVt(vVt.begin() + shift, vVt.begin() + shift + size);
+    opt->vt_->set(tmpVt);
+
+    id++;
+  }
+}
+
+void Adam::save(const std::string& name,
+                std::vector<Ptr<OptimizerBase>> opts,
+                size_t totalSize) {
+  LOG(info, "Saving Adam parameters to {}", name);
+
+  std::vector<float> vMt;
+  std::vector<float> vVt;
+
+  for(auto optBase : opts) {
+    auto opt = std::dynamic_pointer_cast<Adam>(optBase);
+
+    std::vector<float> tmp;
+    opt->mt_->get(tmp);
+    vMt.insert(vMt.end(), tmp.begin(), tmp.end());
+    opt->vt_->get(tmp);
+    vVt.insert(vVt.end(), tmp.begin(), tmp.end());
+  }
+
+  // the shape is the same for mt_ and vt_
+  unsigned* shape = new unsigned[2];
+  shape[0] = 1;
+  shape[1] = vMt.size();
+
+  cnpy::npz_save(name, "adam_mt", vMt.data(), shape, 2, "w");
+  cnpy::npz_save(name, "adam_vt", vVt.data(), shape, 2, "a");
+
+  delete[] shape;
 }
 
 void Adam::resetStats() {