+++ src/translator/translator.h
@@ -32,25 +32,17 @@ public:
     auto vocabs = options_->get<std::vector<std::string>>("vocabs");
     trgVocab_->load(vocabs.back());
 
-    auto devices = options_->get<std::vector<size_t>>("devices");
-
-    DeviceType type = DeviceType::gpu;
-    if(options_->get<size_t>("cpu-threads") > 0) {
-      type = DeviceType::cpu;
-      devices.resize(options_->get<size_t>("cpu-threads"));
-      for(size_t i = 0; i < options_->get<size_t>("cpu-threads"); ++i)
-        devices[i] = i;
-    }
+    auto devices = options_->get<std::vector<DeviceId>>("devices");
 
     ThreadPool threadPool(devices.size(), devices.size());
-
     scorers_.resize(devices.size());
     graphs_.resize(devices.size());
+    
     size_t id = 0;
-    for(size_t device : devices) {
-      auto task = [&](size_t device, size_t id) {
+    for(auto device : devices) {
+      auto task = [&](DeviceId device, size_t id) {
         auto graph = New<ExpressionGraph>(true);
-        graph->setDevice({device, type});
+        graph->setDevice(device);
         graph->reserveWorkspaceMB(options_->get<size_t>("workspace"));
         graphs_[id] = graph;
 
@@ -68,16 +60,8 @@ public:
   void run() {
     data::BatchGenerator<data::Corpus> bg(corpus_, options_);
 
-    // @TODO: ugly
-    auto devices = options_->get<std::vector<size_t>>("devices");
-    DeviceType type = DeviceType::gpu;
-    if(options_->get<size_t>("cpu-threads") > 0) {
-      type = DeviceType::cpu;
-      devices.resize(options_->get<size_t>("cpu-threads"));
-      for(size_t i = 0; i < options_->get<size_t>("cpu-threads"); ++i)
-        devices[i] = i;
-    }
-
+    auto devices = options_->get<std::vector<DeviceId>>("devices");
+    
     ThreadPool threadPool(devices.size(), devices.size());
 
     size_t batchId = 0;
@@ -125,7 +109,7 @@ private:
   std::vector<Ptr<ExpressionGraph>> graphs_;
   std::vector<std::vector<Ptr<Scorer>>> scorers_;
 
-  std::vector<size_t> devices_;
+  std::vector<DeviceId> devices_;
   std::vector<Ptr<Vocab>> srcVocabs_;
   Ptr<Vocab> trgVocab_;
 
@@ -134,7 +118,7 @@ public:
 
   TranslateServiceMultiGPU(Ptr<Config> options)
       : options_(options),
-        devices_(options_->get<std::vector<size_t>>("devices")),
+        devices_(options_->get<std::vector<DeviceId>>("devices")),
         trgVocab_(New<Vocab>()) {
     init();
   }
@@ -150,18 +134,10 @@ public:
     }
     trgVocab_->load(vocabPaths.back());
 
-    DeviceType type = DeviceType::gpu;
-    if(options_->get<size_t>("cpu-threads") > 0) {
-      type = DeviceType::cpu;
-      devices_.resize(options_->get<size_t>("cpu-threads"));
-      for(size_t i = 0; i < options_->get<size_t>("cpu-threads"); ++i)
-        devices_[i] = i;
-    }
-
     // initialize scorers
-    for(auto& device : devices_) {
+    for(auto device : devices_) {
       auto graph = New<ExpressionGraph>(true);
-      graph->setDevice({device, type});
+      graph->setDevice(device);
       graph->reserveWorkspaceMB(options_->get<size_t>("workspace"));
       graphs_.push_back(graph);